Diverge legacy and new backstore to reflect layering change
This commit adds a linear device in between the thin pool and linear
device containing all of the data devices. This will later be used when
the cryptsetup layer is moved on top of the cap device. This way, the
backing device for the crypt device will never change. Changing the
backing device for the dm-crypt device outside of cryptsetup could
result in future problems.